storage: temporary refactoring
