+++ azurerm/internal/services/storage/id.go
@@ -7,13 +7,6 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 )
 
-type AccountID struct {
-	Name          string
-	ResourceGroup string
-
-	ID azure.ResourceID
-}
-
 func AccountIDSchema() *schema.Schema {
 	return &schema.Schema{
 		Type:         schema.TypeString,
@@ -21,30 +14,6 @@ func AccountIDSchema() *schema.Schema {
 		ForceNew:     true,
 		ValidateFunc: ValidateAccountID,
 	}
-}
-
-func ParseAccountID(id string) (*AccountID, error) {
-	storageID, err := azure.ParseAzureResourceID(id)
-	if err != nil {
-		return nil, err
-	}
-
-	resourceGroup := storageID.ResourceGroup
-	if resourceGroup == "" {
-		return nil, fmt.Errorf("%q is missing a Resource Group", id)
-	}
-
-	storageAccountName := storageID.Path["storageAccounts"]
-	if storageAccountName == "" {
-		return nil, fmt.Errorf("%q is missing the `storageAccounts` segment", id)
-	}
-
-	accountId := AccountID{
-		Name:          storageAccountName,
-		ResourceGroup: resourceGroup,
-		ID:            *storageID,
-	}
-	return &accountId, nil
 }
 
 func ValidateAccountID(i interface{}, k string) (warnings []string, errors []error) {
