+++ azurerm/internal/services/storage/parsers/id_test.go
@@ -1,4 +1,4 @@
-package storage
+package parsers
 
 import (
 	"testing"
