+++ azurerm/resource_arm_storage_data_lake_gen2_filesystem.go
@@ -11,6 +11,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/features"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parsers"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 	"github.com/tombuildsstuff/giovanni/storage/2018-11-09/datalakestore/filesystems"
@@ -77,7 +78,7 @@ func resourceArmStorageDataLakeGen2FileSystemCreate(d *schema.ResourceData, meta
 	ctx, cancel := timeouts.ForCreate(meta.(*ArmClient).StopContext, d)
 	defer cancel()
 
-	storageID, err := storage.ParseAccountID(d.Get("storage_account_id").(string))
+	storageID, err := parsers.ParseAccountID(d.Get("storage_account_id").(string))
 	if err != nil {
 		return err
 	}
@@ -134,7 +135,7 @@ func resourceArmStorageDataLakeGen2FileSystemUpdate(d *schema.ResourceData, meta
 		return err
 	}
 
-	storageID, err := storage.ParseAccountID(d.Get("storage_account_id").(string))
+	storageID, err := parsers.ParseAccountID(d.Get("storage_account_id").(string))
 	if err != nil {
 		return err
 	}
